Clean up unused NullOptimization and move test module out of test file
- Remove unused _no_optimization / NullOptimization from SingleModuleStep,
  SeparateRadiationStep, FCN3Step, and SeparatedModuleStep
- Move SimpleSeparatedModule and its builder into separated_module.py with
  deferred registration via register_test_types(), eliminating test-to-test
  import dependency
- Fix SeparatedModuleConfig.build docstring to document labels parameter
