+++ fme/core/step/radiation.py
@@ -15,7 +15,6 @@ from fme.core.device import get_device
 from fme.core.distributed import Distributed
 from fme.core.normalizer import NetworkAndLossNormalizationConfig, StandardNormalizer
 from fme.core.ocean import Ocean, OceanConfig
-from fme.core.optimization import NullOptimization
 from fme.core.packer import Packer
 from fme.core.registry import CorrectorSelector, ModuleSelector
 from fme.core.step.args import StepArgs
@@ -293,7 +292,6 @@ class SeparateRadiationStep(StepABC):
         self.radiation_module = radiation_module.to(get_device())
         self._img_shape = dataset_info.img_shape
         self._config = config
-        self._no_optimization = NullOptimization()
 
         init_weights(self.modules)
         dist = Distributed.get_instance()
