+++ fme/core/step/separated_module.py
@@ -14,7 +14,6 @@ from fme.core.device import get_device
 from fme.core.distributed import Distributed
 from fme.core.normalizer import NetworkAndLossNormalizationConfig, StandardNormalizer
 from fme.core.ocean import Ocean, OceanConfig
-from fme.core.optimization import NullOptimization
 from fme.core.packer import Packer
 from fme.core.registry import CorrectorSelector, SeparatedModuleSelector
 from fme.core.step.args import StepArgs
@@ -275,7 +274,6 @@ class SeparatedModuleStep(StepABC):
         init_weights(self.modules)
         self._img_shape = dataset_info.img_shape
         self._config = config
-        self._no_optimization = NullOptimization()
 
         self.module = self.module.wrap_module(dist.wrap_module)
         self.secondary_decoder = self.secondary_decoder.wrap_module(dist.wrap_module)
